fix(mcp): preserve typed runtime seam boundary
